Level: Implemented threaded chunk I/O
