+++ src/pocketmine/level/format/io/region/AnvilChunkProvider.php
@@ -1,0 +1,145 @@
+<?php
+
+/*
+ *
+ *  ____            _        _   __  __ _                  __  __ ____
+ * |  _ \ ___   ___| | _____| |_|  \/  (_)_ __   ___      |  \/  |  _ \
+ * | |_) / _ \ / __| |/ / _ \ __| |\/| | | '_ \ / _ \_____| |\/| | |_) |
+ * |  __/ (_) | (__|   <  __/ |_| |  | | | | | |  __/_____| |  | |  __/
+ * |_|   \___/ \___|_|\_\___|\__|_|  |_|_|_| |_|\___|     |_|  |_|_|
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * @author PocketMine Team
+ * @link http://www.pocketmine.net/
+ *
+ *
+*/
+
+declare(strict_types=1);
+
+namespace pocketmine\level\format\io\region;
+
+use pocketmine\level\format\Chunk;
+use pocketmine\level\format\ChunkException;
+use pocketmine\level\format\io\ChunkUtils;
+use pocketmine\level\format\SubChunk;
+use pocketmine\nbt\BigEndianNBTStream;
+use pocketmine\nbt\NBT;
+use pocketmine\nbt\tag\ByteArrayTag;
+use pocketmine\nbt\tag\CompoundTag;
+use pocketmine\nbt\tag\IntArrayTag;
+use pocketmine\nbt\tag\ListTag;
+use pocketmine\utils\MainLogger;
+
+class AnvilChunkProvider extends McRegionChunkProvider{
+
+	protected function getRegionFileExtension() : string{
+		return "mca";
+	}
+
+	protected function nbtSerialize(Chunk $chunk) : string{
+		$nbt = new CompoundTag("Level", []);
+		$nbt->setInt("xPos", $chunk->getX());
+		$nbt->setInt("zPos", $chunk->getZ());
+
+		$nbt->setByte("V", 1);
+		$nbt->setLong("LastUpdate", 0); //TODO
+		$nbt->setLong("InhabitedTime", 0); //TODO
+		$nbt->setByte("TerrainPopulated", $chunk->isPopulated() ? 1 : 0);
+		$nbt->setByte("LightPopulated", $chunk->isLightPopulated() ? 1 : 0);
+
+		$subChunks = [];
+		foreach($chunk->getSubChunks() as $y => $subChunk){
+			if($subChunk->isEmpty()){
+				continue;
+			}
+
+			$tag = $this->serializeSubChunk($subChunk);
+			$tag->setByte("Y", $y);
+			$subChunks[] = $tag;
+		}
+		$nbt->setTag(new ListTag("Sections", $subChunks, NBT::TAG_Compound));
+
+		$nbt->setByteArray("Biomes", $chunk->getBiomeIdArray());
+		$nbt->setIntArray("HeightMap", $chunk->getHeightMapArray());
+
+		$nbt->setTag(new ListTag("Entities", $chunk->NBTentities, NBT::TAG_Compound));
+		$nbt->setTag(new ListTag("TileEntities", $chunk->NBTtiles, NBT::TAG_Compound));
+
+		//TODO: TileTicks
+
+		$writer = new BigEndianNBTStream();
+		$nbt->setName("Level");
+		$writer->setData(new CompoundTag("", [$nbt]));
+
+		return $writer->writeCompressed(ZLIB_ENCODING_DEFLATE, RegionLoader::$COMPRESSION_LEVEL);
+	}
+
+	protected function serializeSubChunk(SubChunk $subChunk) : CompoundTag{
+		return new CompoundTag("", [
+			new ByteArrayTag("Blocks", ChunkUtils::reorderByteArray($subChunk->getBlockIdArray())), //Generic in-memory chunks are currently always XZY
+			new ByteArrayTag("Data", ChunkUtils::reorderNibbleArray($subChunk->getBlockDataArray())),
+			new ByteArrayTag("SkyLight", ChunkUtils::reorderNibbleArray($subChunk->getBlockSkyLightArray(), "\xff")),
+			new ByteArrayTag("BlockLight", ChunkUtils::reorderNibbleArray($subChunk->getBlockLightArray()))
+		]);
+	}
+
+	protected function nbtDeserialize(string $data){
+		$nbt = new BigEndianNBTStream();
+		try{
+			$nbt->readCompressed($data);
+
+			$chunk = $nbt->getData()->getCompoundTag("Level");
+
+			if($chunk === null){
+				throw new ChunkException("Invalid NBT format");
+			}
+
+			$subChunks = [];
+			$subChunksTag = $chunk->getListTag("Sections") ?? [];
+			foreach($subChunksTag as $subChunk){
+				if($subChunk instanceof CompoundTag){
+					$subChunks[$subChunk->getByte("Y")] = $this->deserializeSubChunk($subChunk);
+				}
+			}
+
+			if($chunk->hasTag("BiomeColors", IntArrayTag::class)){
+				$biomeIds = ChunkUtils::convertBiomeColors($chunk->getIntArray("BiomeColors")); //Convert back to original format
+			}else{
+				$biomeIds = $chunk->getByteArray("Biomes", "", true);
+			}
+
+			$result = new Chunk(
+				$chunk->getInt("xPos"),
+				$chunk->getInt("zPos"),
+				$subChunks,
+				$chunk->hasTag("Entities", ListTag::class) ? $chunk->getListTag("Entities")->getValue() : [],
+				$chunk->hasTag("TileEntities", ListTag::class) ? $chunk->getListTag("TileEntities")->getValue() : [],
+				$biomeIds,
+				$chunk->getIntArray("HeightMap", [])
+			);
+			$result->setLightPopulated($chunk->getByte("LightPopulated", 0) !== 0);
+			$result->setPopulated($chunk->getByte("TerrainPopulated", 0) !== 0);
+			$result->setGenerated();
+
+			return $result;
+		}catch(\Throwable $e){
+			MainLogger::getLogger()->logException($e);
+
+			return null;
+		}
+	}
+
+	protected function deserializeSubChunk(CompoundTag $subChunk) : SubChunk{
+		return new SubChunk(
+			ChunkUtils::reorderByteArray($subChunk->getByteArray("Blocks")),
+			ChunkUtils::reorderNibbleArray($subChunk->getByteArray("Data")),
+			ChunkUtils::reorderNibbleArray($subChunk->getByteArray("SkyLight"), "\xff"),
+			ChunkUtils::reorderNibbleArray($subChunk->getByteArray("BlockLight"))
+		);
+	}
+}
